extract CLI template path auto-detection into prompt-templates core
